+++ src/components/common/Header.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Header = () => {
+  return <div>헤더라능</div>;
+};
+
+export default Header;
