+++ src/components/common/Layout.jsx
@@ -1,7 +1,13 @@
 import React from "react";
+import Header from "../common/Header";
 
-const Layout = () => {
-  return <></>;
+const Layout = ({ children }) => {
+  return (
+    <div>
+      <Header />
+      {children}
+    </div>
+  );
 };
 
 export default Layout;
