+++ src/components/component/Que.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Que = () => {
+  return <div>Que</div>;
+};
+
+export default Que;
