+++ src/css_modules/JoinPage.module.css
@@ -1,5 +1,3 @@
-@import url(../fonts/fonts.css);
-
 .joinWarp {
   width: 300px;
   padding: 30px;
@@ -7,7 +5,7 @@
   text-align: center;
   border: solid 3px black;
   border-radius: 20px;
-  font-family: 'Noto Sans KR', sans-serif;
+  font-family: "Noto Sans KR", sans-serif;
 }
 
 .joinWarp h2 {
@@ -22,14 +20,14 @@
   width: calc(100% - 20px);
   margin-bottom: 20px;
   border: 2px solid black;
-  font-family: 'Noto Sans KR', sans-serif;
+  font-family: "Noto Sans KR", sans-serif;
   font-weight: 700;
   border-radius: 15px;
 }
 
 .joinWarp input::placeholder {
   color: black;
-  font-family: 'Noto Sans KR', sans-serif;
+  font-family: "Noto Sans KR", sans-serif;
   font-weight: 700;
 }
